+++ History.md
@@ -1,3 +1,15 @@
+# Version 0.8.7
+
+Release date: 2012-04-3
+
+### Added
+
+* FindFrame finds both iframes and old school frameset frames, FindIframe is now obsolete.
+
+### Fixed
+
+* Location respects window scope in all cases
+
 # Version 0.8.6
 
 Release date: 2012-04-18